# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Zoo API
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local developement server
...
paths:
...
  /api/user:
    get:
      summary: Get all user
      description: Gets all user. Administrators Only 
      security:
        - bearerAuth: []
      tags:
        - User
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/NotPermitted'
        '429':
          $ref: '#/components/responses/RateLimitHit'

    post:
      summary: Create an user
      description: Creates a user using the provided data. Administrators Only
      security:
        - bearerAuth: []
      tags:
        - User
      requestBody:
        description: Create user data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/NotPermitted'
        '400':
          $ref: '#/components/responses/InvalidUserInput'
        '409':
          $ref: '#/components/responses/EmailConflict'
        '429':
          $ref: '#/components/responses/RateLimitHit'
...
  /api/zoo:
    get:
      tags:
        - Zoo
      summary: Get all zoo
      description: Retrieve all zoo data from the database
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully retrieved data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Zoo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitHit'

    post:
      tags:
        - Zoo
      security:
        - bearerAuth: []
      summary: Create a zoo
      description: Create a zoo with provided JSON body
      requestBody:
        required: true
        description: The data to create a zoo from
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Zoo'
      responses:
        '201':
          description: Zoo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Zoo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/InvalidZooInput'
        '409':
          $ref: '#/components/responses/ZooConflict'
        '429':
          $ref: '#/components/responses/RateLimitHit'
...
components:
  schemas:
    Zoo:
      type: object
      properties:
        id:
          readOnly: true
          type: integer
          example: 1
        name:
          type: string
          example: 'Green Zoo'
        location:
          type: string
          example: 'Charlotte, NC'
        yearOpened:
          type: integer
          example: 2026
        ownerId:
          type: integer
          example: 1
...
    User:
      type: object
      properties:
        id:
          readOnly: true
          type: integer
          example: 1
        email:
          type: string
          example: 'example@example.com'
        password:
          writeOnly: true
          type: string
          example: 'password123'
        role:
          $ref: '#/components/schemas/Role'
          example: 'USER'
          readOnly: true
...
    Role:
      type: string
      enum:
        - 'USER'
        - 'ADMIN'
...
  responses:
...
    Unauthorized:
      description: Client is not authenticated
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
            example:
              error: 'Not authenticated. Please provide a valid token.'
    InvalidUserInput:
      description: Bad request due to invalid data
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  type: string
            example:
              errors:
                - "Field 'email' is missing"
                - "Field 'password' is missing"
    InvalidZooInput:
      description: Bad request due to invalid data
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  type: string
            example:
              errors:
                - "Field 'name' is missing"
                - "Field 'location' is missing"
    NotPermitted:
      description: Client is not authorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
            example:
              error: 'Forbidden: insufficient permission.'
    RateLimitHit:
      description: The client has tried to make too many connections and the rate limiter was hit
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
            example:
              error: 'Too many requests. Try again later'
    EmailConflict:
      description: Email already in use
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
            example:
              error: 'Conflict: Email is already in use'
    ZooConflict:
      description: Zoo creation/updating had conflicts
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
            example:
              error: 'Conflict: owner already has a zoo with this name in this location'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT